moved stuff around and cleaned up the create blog form
