+++ src/components/ui/context-controller/context-controller.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { useGetSections } from "@/hooks/hooks";
+import { useGetSections } from "@/lib/hooks/hooks";
 import BlogComponent from "../blog-component/blog-component";
 import EditBlogComponent from "../sections/edit-section-components/edit-blog-component/edit-blog-component"
 import { useState } from "react";
