+++ src/components/ui/display-blogs/display-blogs.tsx
@@ -11,36 +11,36 @@ export function DisplayBlogs({ blogsPromise }: { blogsPromise: Promise<ResponseD
 
     const result = use(blogsPromise);
 
-    if ('error' in result) return < NoticeDisplay>Error: {result.error}</ NoticeDisplay>;
+    if ('error' in result) return <NoticeDisplay>Error: {result.error}</NoticeDisplay>;
 
     const { blogs, hasMore } = result;
 
-    if (!blogs || blogs.length === 0) return < NoticeDisplay>No blogs found</ NoticeDisplay>;
+    if (!blogs || blogs.length === 0) return <NoticeDisplay>No blogs found</NoticeDisplay>;
 
     const publishedBlogs = blogs.filter(b => b.published);
     const unpublishedBlogs = blogs.filter(b => !b.published);
 
     return (
-        <div className="flex flex-col justify-center items-center gap-4  p-4 w-full lg:w-4/6 sm:w-5/6">
-            {unpublishedBlogs.length > 0 && (
-                <>
-                    <h2 className="text-2xl">Unpublished Drafts</h2>
-                    <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4 w-full">
-                        {unpublishedBlogs.map((blog: BlogData) => (
-                            <BlogCard key={blog.id} blog={blog} />
-                        ))}
-                    </div>
-                </>
-            )}
+        <div className="flex flex-col justify-center items-center gap-4 p-4 w-full">
             {publishedBlogs.length > 0 && (
-                <>
+                <div className="flex flex-col items-center w-full gap-4">
                     <h2 className="text-2xl">Published Blogs</h2>
-                    <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4 w-full">
+                    <div className={`flex flex-wrap gap-4 w-full ${publishedBlogs.length === 1 ? 'justify-center' : 'justify-start'}`}>
                         {publishedBlogs.map((blog: BlogData) => (
                             <BlogCard key={blog.id} blog={blog} />
                         ))}
                     </div>
-                </>
+                </div>
+            )}
+            {unpublishedBlogs.length > 0 && (
+                <div className="flex flex-col items-center w-full gap-4">
+                    <h2 className="text-2xl">Unpublished Drafts</h2>
+                    <div className="flex flex-wrap gap-4 w-full justify-start">
+                        {unpublishedBlogs.map((blog: BlogData) => (
+                            <BlogCard key={blog.id} blog={blog} />
+                        ))}
+                    </div>
+                </div>
             )}
             <PaginationButtons page={1} path={"/"} hasMoreBlogs={hasMore} />
         </div>
@@ -49,7 +49,7 @@ export function DisplayBlogs({ blogsPromise }: { blogsPromise: Promise<ResponseD
 
 function BlogCard({ blog }: { blog: BlogData }) {
     return (
-        <div className="flex flex-col bg-[linear-gradient(to_bottom_right,var(--primary),var(--secondary))] rounded-sm shadow-xl gap-4 pt-2 justify-center items-center w-full" >
+        <div className="flex flex-col bg-[linear-gradient(to_bottom_right,var(--primary),var(--secondary))] rounded-sm shadow-xl gap-4 pt-2 justify-center items-center flex-grow flex-shrink-0 basis-80 max-w-sm" >
             <h1 className="text-2xl text-center p-2">{blog.title}</h1>
             {blog.image_src &&
                 <Image
