+++ src/components/ui/sections/add-section-components/add-section-form/add-section-form.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { useState } from "react"
 import { TitleSection } from "../title-section"
-import { useAddSection } from "@/mutations/mutations";
+import { useAddSection } from "@/lib/mutations/mutations";
 import ParagraphSection from "../paragraph-section";
 import CodeSection from "../code-section";
 import { ImageSection } from "../image-section";
